Rozhrani/requirements.txt
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
idna==3.10
influxdb-client==1.47.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
paho-mqtt==2.1.0
pycparser==2.22
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reactivex==4.0.4
requests==2.32.3
retrying==1.3.4
setuptools==75.6.0
six==1.17.0
tinytuya==1.15.1
tuya-connector-python==0.1.2
tuya-iot-py-sdk==0.6.6
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.1.3
